refactor: Extract the state-indexer common logic into a separate crate. Convert state-indexer to use the new library
