+++ logic-state-indexer/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "state-indexer"
+name = "logic-state-indexer"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
@@ -29,14 +29,9 @@ tokio-stream = "0.1"
 tracing = "0.1.34"
 
 configuration.workspace = true
-database.workspace = true
+database-new.workspace = true
 epoch-indexer.workspace = true
-logic-state-indexer.workspace = true
 
 near-primitives.workspace = true
 near-indexer-primitives.workspace = true
 near-jsonrpc-client.workspace = true
-near-lake-framework.workspace = true
-
-[features]
-tracing-instrumentation = ["configuration/tracing-instrumentation"]
